feat: display completions in fzf
Display a list of completions in fzf (5 by default, but the number is configurable using the 'completions' option in fish-ai.ini) when pressing Ctrl+Space.

Completions are loading in the background and the user can cycle through the available completions using the arrow keys (up and down) and press Enter to make a choice. The selected completion is then copied to the command line buffer.

Resolves: #18
Signed-off-by: Bastian Fredriksson <bastian.fredriksson@keyfactor.com>