feat: add commandline history to completion prompt
Send the output from history search --max {N} --prefix {cmd_buffer} to the LLM when completing a command. This can be used to personalise the suggestions created by the LLM.

N is configurable and disabled by default (N = 0), since the history may contain sensitive information and users should explicitly opt-in.