fix: decrease max_tokens to 1024
The previous value of 8192 is too large and causes the following error: "max_tokens is too large: 8192. This model supports at most 4096 completion tokens, whereas you provided 8192."

max_tokens indicates the number of tokens in the *output* and input + max_tokens < context length for model. 1024 should be a safe value to use and sufficiently large for creating even the most advanced CLI commands.

Resolves: #28