feat: use os truststore on linux systems
If the user is sending requests to a server with a TLS server certificate signed by a private CA (e.g. because the server is hosted on a company intranet), connections will fail unless the user is manually trusting the correcy CA certificate.

Attempt to use the OS truststore, to not throw people into the PKI rabbithole, by symlinking to the certificate bundle in /etc/ssl or /etc/pki. Should work on Fedora, Arch Linux and Ubuntu. macOS stores its OS truststore in a keychain, not as a plain PEM on the filesystem, so it will probably not work.

Fallback will be the Mozilla truststore provided by the certifi package.