fix: use ctrl-space on macos and -k nul on linux
Ctrl+Space is called different things on different
systems. This should make it work on both macOS
and Linux out of the box. Also, make it possible
to override the default key bindings (Ctrl+P and
Ctrl+Space with custom ones, using environment
variables).

Thank you to amir20 for doing the research. 🤝

Closes: #183