fix: use correct thinking config for gemini
The plugin does not work when the Google provider is used, unless a model which supports thinking level (such as gemini-3-flash-preview) is used. The default option is gemini-2.5-flash-lite which uses thinking budget parameter.

Providing the appropriate configuration depending on which model is being used.

New models (probably called gemini-4-something) will use not a thinking config, and it may break, but we'll deal with that when that happens.

Closes: #476